migrated bench to main command
