+++ Cargo.toml
@@ -2,17 +2,16 @@
 resolver = "2"
 
 members = [
-  "bench",
-  "better_io",
-  "pco",
-  "pco_c",
-  "pco_cli",
-  "pco_python",
-  "quantile-compression/ffi",
-  "quantile-compression/q_compress",
+    "better_io",
+    "pco",
+    "pco_c",
+    "pco_cli",
+    "pco_python",
+    "quantile-compression/ffi",
+    "quantile-compression/q_compress",
 ]
 
 exclude = [
-  # conflicting dependencies and not worth upgrading
-  "quantile-compression/q_compress_cli"
+    # conflicting dependencies and not worth upgrading
+    "quantile-compression/q_compress_cli"
 ]
