+++ pco_cli/Cargo.toml
@@ -19,9 +19,15 @@ path = "src/main.rs"
 anyhow = "1.0.71"
 arrow = { version = "49.0.0", features = ["csv"], default-features = false }
 better_io = { version = "0.0.0", path = "../better_io" }
+blosc-src = { version = "0.3", features = ["zstd"] }
+bytes = "1.1.0"
 clap = { version = "4.5.3", features = ["derive"] }
 parquet = { version = "49.0.0", features = ["arrow", "base64", "snap", "zstd"], default-features = false }
 pco = { version = "0.2", path = "../pco" }
+q_compress = { path = "../quantile-compression/q_compress" }
+snap = "1.1.0"
 toml = "0.8.12"
 tabled = "0.15.0"
 serde = { version = "1.0.197", features = ["derive"] }
+zstd = "0.13"
+zstd-sys = "2.0.9"
