+++ pco_cli/src/bench/codecs/pco.rs
@@ -2,8 +2,8 @@ use anyhow::{anyhow, Result};
 
 use pco::{FloatMultSpec, IntMultSpec, PagingSpec};
 
-use crate::codecs::CodecInternal;
-use crate::dtypes::Dtype;
+use crate::bench::codecs::CodecInternal;
+use crate::bench::dtypes::Dtype;
 
 #[derive(Clone, Debug, Default)]
 pub struct PcoConfig {
