+++ pco_cli/src/bench/codecs/qco.rs
@@ -1,7 +1,7 @@
 use anyhow::{anyhow, Result};
 
-use crate::codecs::CodecInternal;
-use crate::dtypes::Dtype;
+use crate::bench::codecs::CodecInternal;
+use crate::bench::dtypes::Dtype;
 
 #[derive(Clone, Debug, Default)]
 pub struct QcoConfig {
