+++ pco_cli/src/bench/codecs/snappy.rs
@@ -3,8 +3,8 @@ use std::io::{Read, Write};
 
 use anyhow::{anyhow, Result};
 
-use crate::codecs::{utils, CodecInternal};
-use crate::dtypes::Dtype;
+use crate::bench::codecs::{utils, CodecInternal};
+use crate::bench::dtypes::Dtype;
 
 #[derive(Clone, Debug, Default)]
 pub struct SnappyConfig {}
