+++ pco_cli/src/main.rs
@@ -4,17 +4,20 @@ use clap::Parser;
 use crate::opt::{Opt, OptWrapper};
 
 mod arrow_handlers;
+mod bench;
 mod compress;
 mod core_handlers;
 mod decompress;
 mod dtypes;
 mod inspect;
 mod opt;
+mod parse;
 mod utils;
 
 fn main() -> Result<()> {
   let opt = OptWrapper::parse().opt;
   match opt {
+    Opt::Bench(bench_opt) => bench::bench(bench_opt)?,
     Opt::Compress(compress_opt) => compress::compress(compress_opt)?,
     Opt::Decompress(decompress_opt) => decompress::decompress(decompress_opt)?,
     Opt::Inspect(inspect_opt) => inspect::inspect(inspect_opt)?,
